requirements.txt
alembic==1.7.7
aniso8601==9.0.1
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
colorama==0.4.4
coverage==6.3.2
Flask==2.1.1
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==4.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
nose2==0.11.0
passlib==1.7.4
pdfkit==1.0.0
Pillow==9.1.0
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.34
Werkzeug==2.0.3
zipp==3.7.0
